This is a helper method to load virtualenvs quickly.
The function `use_env` is added to the autoenv shell script, so it is
sourced and available to be used inside `.env` scripts.  This makes it
easy to declare what virtualenv should be loaded or created when
switching to a specific dir.

For example:

    use_env myenv

The function is cheap when the environment is already active and will
only invoke the expensive `workon` and `mkvirtualenv` commands when
absolutely necessary.