Updating version for the next PyPI release.
A dev release was published to PyPI as v1.0.0 back in 2012, making pip
prioritize it over newer releases (with new features). After the next
PyPI upload, `pip install autoenv` and `pip install -U autoenv` will
both work as expected.

This is not ideal semantic versioning, but it follows the rules.