autoenv_leave: Only match prefix on path boundaries
Suppose we `cd` from `a/b` to `a/bz`.

Expected behavior: Invoke `a/b/.env.leave.

Actual behavior prior to this commit: No env-leave is invoked.

The issue is that we are naively checking for string prefixes in
`autoenv_leave` in order to determine whether one directory is an
ancestor of another. This commit makes the checking slightly less naive.